create western module
