+++ spec/zakuro/calculation/summary/western/yaml/range.yaml
