+++ spec/zakuro/calculation/summary/western/yaml/single.yaml
